% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_def.R
\name{auc.imbDis}
\alias{auc.imbDis}
\title{Area Under the Receiver Operator Curve (ROC AUC)}
\usage{
\method{auc}{imbDis}(obj)
}
\arguments{
\item{obj}{An imbDis object.}
}
\value{
data.frame with columns: bins, AUC, sample size.
}
\description{
Calculates the Area under the Receiver Operator Curve (ROC AUC) for each bin frequency 
with the same sample size.
}
