% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{brier}
\alias{brier}
\title{Brier Score}
\usage{
brier(obj)
}
\arguments{
\item{obj}{an object of class inheriting from "imbDis".}
}
\value{
data.frame with columns: bins, Brier score, sample size.
}
\description{
Calculates the Brier score for each bin frequency 
with the same sample size.
}
