% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_def.R
\name{.calc_samplesize}
\alias{.calc_samplesize}
\title{Calculates the sample size}
\usage{
.calc_samplesize(labels_01, bins)
}
\arguments{
\item{labels}{Vector of ground-truth labels.}

\item{case}{An integer or string that is the case class. Must be a value in labels.}
}
\value{
Sample size integer.
}
\description{
Helper function to convert labels vector to 0-1 labels. 1 corresponds to the case label.
}
