% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_def.R
\name{.param_check}
\alias{.param_check}
\title{Parameter Check}
\usage{
.param_check(labels, pred, case, bins)
}
\arguments{
\item{labels}{Vector of ground-truth labels}

\item{pred}{Vector of predicted probabilities}

\item{case}{An integer or string that is the case class. Must be a value in labels.}

\item{bins}{Vector of different class imbalance frequencies. Default is seq(0.05, 0.5, 0.05)}
}
\description{
Helper function to ensure parameters are valid.
}
