% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_def.R
\name{.standardiseLabels}
\alias{.standardiseLabels}
\title{Standardised Labels}
\usage{
.standardiseLabels(labels, case)
}
\arguments{
\item{labels}{Vector of ground-truth labels.}

\item{case}{An integer or string that is the case class. Must be a value in labels.}
}
\value{
Standardised labels vector.
}
\description{
Helper function to convert labels vector to 0-1 labels. 1 corresponds to the case label.
}
