% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_def.R
\name{imbDis}
\alias{imbDis}
\title{imbDis - Imbalanced Discrimination}
\usage{
imbDis(labels, pred, case, bins = seq(0.05, 0.5, 0.05))
}
\arguments{
\item{labels}{Vector of ground-truth labels.}

\item{pred}{Vector of predicted probabilities.}

\item{case}{An integer or string that is the case class. Must be a value in labels.}

\item{bins}{Vector of different class imbalance frequencies. Default is seq(0.05, 0.5, 0.05).}
}
\value{
A simMetric object - S3 class that contains the input parameters, standardized labels, 
    and sample size.
}
\description{
This function creates an object of type \code{imbDis}, an S3 Class. 
imbDis objects can be passed into the associated methods as arguments.
Examples of associated methods are \code{auc}, \code{brier} and \code{logLoss}.
}
