% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{logLoss.imbDis}
\alias{logLoss.imbDis}
\title{Log Loss/Binary Cross-Entropy}
\usage{
\method{logLoss}{imbDis}(obj)
}
\arguments{
\item{obj}{an object of class inheriting from "imbDis".}
}
\value{
data.frame with columns: bins, log loss, sample size.
}
\description{
Calculates the log loss for each bin frequency 
with the same sample size.
}
