% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{manualMetric}
\alias{manualMetric}
\title{Other metrics framework}
\usage{
manualMetric(obj, f)
}
\arguments{
\item{obj}{an object of class inheriting from "imbDis".}

\item{f}{A function that must be of the format f(labels, preds).}
}
\value{
data.frame with columns: bins, metric, sample size.
}
\description{
Calculates a user-input metric for each bin frequency with the same sample size.
Allows imbDis metric calculation using the predicted labels instead of probabilities.
}
