% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imbDis.R
\name{imbDis}
\alias{imbDis}
\title{imbDis - Imbalanced Discrimination}
\usage{
imbDis(labels, pred, case, bins = seq(0.05, 0.5, 0.05))
}
\arguments{
\item{labels}{Vector of ground-truth labels.}

\item{pred}{Vector of predicted probabilities.}

\item{case}{A value present in the \code{labels} argument, that denotes the case class label.}

\item{bins}{Vector of class imbalance frequencies. Values must be larger than 0 and smaller than 1. Default is seq(0.05, 0.5, 0.05).}
}
\value{
A simMetric object - S3 class that contains the input parameters, standardized labels, 
    and sample size.
}
\description{
This function creates an object of type \code{imbDis}, an S3 Class. 
imbDis objects can be passed into the associated methods as arguments.
Examples of associated methods are \code{auc}, \code{brier} and \code{logLoss}.
}
\examples{
# using the mtcars dataset
LR = glm(vs ~ mpg + hp + wt, data = mtcars, family = "binomial")
preds = predict(LR, mtcars, type = "response")  # equiv. to LR$fitted.values

# define the imbDis object
imbD = imbDis(mtcars$vs, preds, 1, seq(0.1, 0.9, 0.1))

# metric results
auc(imbD)
brier(imbD)
logLoss(imbD)

}
